gui/requirements.txt
alabaster==0.7.12
astroid==2.3.3
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
docutils==0.16
idna==2.9
imagesize==1.2.0
isort==4.3.21
Jinja2==2.11.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.2
packaging==20.3
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
snowballstemmer==2.0.0
Sphinx==3.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
urllib3==1.25.9
wrapt==1.11.2
